+++ src/components/Synthetics/OrderList/OrderList.tsx
@@ -8,6 +8,8 @@ import {
   isLimitOrderType,
   isSwapOrderType,
   isTriggerDecreaseOrderType,
+  sortPositionOrders,
+  sortSwapOrders,
 } from "domain/synthetics/orders";
 import { cancelOrdersTxn } from "domain/synthetics/orders/cancelOrdersTxn";
 import { PositionsInfoData } from "domain/synthetics/positions";
@@ -60,19 +62,7 @@ export function OrderList(p: Props) {
       { swapOrders: [] as SwapOrderInfo[], positionOrders: [] as PositionOrderInfo[] }
     );
 
-    const sortedPositionOrders = positionOrders.sort((a, b) => {
-      const nameComparison = a.marketInfo.name.localeCompare(b.marketInfo.name);
-      if (nameComparison !== 0) return nameComparison;
-      return a.triggerPrice.sub(b.triggerPrice).isNegative() ? -1 : 1;
-    });
-
-    const sortedSwapOrders = swapOrders.sort((a, b) => {
-      const collateralComparison = a.targetCollateralToken.symbol.localeCompare(b.targetCollateralToken.symbol);
-      if (collateralComparison !== 0) return collateralComparison;
-      return a.minOutputAmount.sub(b.minOutputAmount).isNegative() ? -1 : 1;
-    });
-
-    return [...sortedPositionOrders, ...sortedSwapOrders];
+    return [...sortPositionOrders(positionOrders), ...sortSwapOrders(swapOrders)];
   }, [p.ordersData]);
 
   const isAllOrdersSelected = orders.length > 0 && orders.every((o) => p.selectedOrdersKeys?.[o.key]);
